requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.1
requests>=2.31.0
pydantic>=2.9.0
supabase>=2.0.0
